+++ example/a.ts
@@ -1,30 +1,7 @@
 import { Elysia, t } from '../src'
 
-const app = new Elysia().get(
-	'/',
-	() => {
-		return [
-			{
-				id: 'testId',
-				date: new Date('2025-07-11T00:00:00.000Z'),
-				name: 'testName',
-				needNormalize: 'yes'
-			}
-		]
-	},
-	{
-		response: {
-			200: t.Array(
-				t.Object({
-					id: t.String(),
-					date: t.Date(),
-					name: t.String()
-				})
-			)
-		}
-	}
-)
+using plugin = new Elysia().get('a', 'a')
 
-app.handle(new Request('http://localhost:3000/'))
-	.then((x) => x.json())
-	.then(console.log)
+const main = new Elysia().use(plugin)
+
+main.listen(3000)